requirements.txt
Flask
numpy
Werkzeug
abc1
Pillow
